sync is more or less finished
