+++ lib/elastic_manager.rb
@@ -8,6 +8,11 @@ require 'elastic_manager/request'
 # require 'elastic_manager/close'
 # require 'elastic_manager/chill'
 require 'elastic_manager/snapshot'
+require 'elastic_manager/sync_roles'
+require 'elastic_manager/sync_users'
+require 'elastic_manager/sync_ilms'
+require 'elastic_manager/sync_templates'
+require 'elastic_manager/sync_spaces'
 # require 'elastic_manager/delete'
 # require 'elastic_manager/snapdelete'
 
@@ -21,277 +26,14 @@ class ElasticManager
   # include Close
   # include Chill
   include Snapshot
+  include Sync
   # include Delete
   # include SnapDelete
-
-  TWIKET_USERS = %w[
-    nginx
-    logstash_internal
-    apm_internal
-  ].freeze
 
   def initialize
     @config  = prepare_config
     @elastic = Request::Elastic.new(@config['system_users']['elastic'])
     @kibana  = Request::Kibana.new(@config['system_users']['elastic'])
-  end
-
-  def sync_roles
-    log.info 'sync roles'
-    elastic_roles = JSON.parse(@elastic.request(:get, '/_xpack/security/role'))
-    roles_for_delete = []
-    elastic_roles.each do |role, params|
-      next if params['metadata']['_reserved']
-      next if role == 'nginx'
-
-      if @config['roles'][role]
-        if @config['roles'][role].key?('retired') && @config['roles'][role]['retired']
-          log.warn "role '#{role}' retired"
-        else
-          next
-        end
-      end
-      roles_for_delete << role
-    end
-    log.info "will delete roles: #{roles_for_delete}"
-
-    roles_for_create = []
-    @config['roles'].each do |role, params|
-      next if params['retired']
-      roles_for_create << role
-    end
-    log.info "will put roles: #{roles_for_create}"
-
-    sleep 3
-
-    roles_for_delete.each do |role|
-      res = @elastic.request(:delete, "/_xpack/security/role/#{role}")
-      log.info "delete #{role}: #{res}"
-    end
-
-    # TODO: (anton.ryabov) mb diff before and put only if changed?
-    roles_for_create.each do |role|
-      res = @elastic.request(:put, "/_xpack/security/role/#{role}", @config['roles'][role]['config'])
-      log.info "create #{role}: #{res}"
-    end
-  end
-
-  def sync_users
-    log.info 'sync users'
-    elastic_users = JSON.parse(@elastic.request(:get, '/_security/user'))
-    users_for_delete = []
-    elastic_users.each do |user, params|
-      next if params['metadata']['_reserved']
-      next if TWIKET_USERS.include?(user)
-
-      if @config['users'][user]
-        if @config['users'][user].key?('retired') && @config['users'][user]['retired']
-          log.warn "user '#{user}' retired"
-        else
-          next
-        end
-      end
-      users_for_delete << user
-    end
-    log.info "will delete users: #{users_for_delete}"
-
-    actual_users     = []
-    users_for_create = []
-    users_for_update = []
-
-    @config['users'].each do |user, params|
-      next if params['retired']
-      actual_users << user
-      if elastic_users[user]
-        if @config['users'][user].key?('roles')
-          next if (@config['users'][user]['roles'] - elastic_users[user]['roles']).length == 0
-          users_for_update << user
-        end
-      else
-        users_for_create << user
-      end
-    end
-    log.info "will update users: #{users_for_update}"
-    log.info "will create users: #{users_for_create}"
-
-    sleep 3
-
-    users_for_delete.each do |user|
-      res = @elastic.request(:delete, "/_security/user/#{user}")
-      log.info "delete #{user}: #{res}"
-    end
-
-    users_for_update.each do |ufu|
-      elastic_users[ufu]['roles'] = (elastic_users[ufu]['roles'] + @config['users'][ufu]['roles']).uniq
-      res = @elastic.request(:put, "/_security/user/#{ufu}", elastic_users[ufu])
-      log.info "update #{ufu}: #{res}"
-    end
-
-    users_for_create.each do |ufc|
-      roles = ['twiket_read_all']
-      # TODO: (anton.ryabov) rewrite this hardcoded cluster check
-      # roles << 'twiket_read_all' # if @config['env'] == 'development'
-      roles = roles + @config['users'][ufc]['roles'] if @config['users'][ufc].key?('roles')
-
-      user_hash = {
-        password: SecureRandom.urlsafe_base64(24),
-        roles:    roles
-      }
-
-      res = @elastic.request(:post, "/_security/user/#{ufc}", user_hash)
-      log.info "create #{ufc}: #{res}"
-    end
-
-    TWIKET_USERS.each do |user|
-      next if elastic_users[user]
-      if @config['system_users'].key?(user) && @config['system_users'][user] != ''
-        if @config['users'].key?(user) && @config['users'][user].key?('roles')
-          user_hash = {
-            password: @config['system_users'][user],
-            roles:    @config['users'][user]['roles']
-          }
-          res = @elastic.request(:post, "/_security/user/#{user}", user_hash)
-          log.info "create #{user}: #{res}"
-        else
-          log.warn "no roles for #{user} in vault"
-        end
-      else
-        log.warn "no password for #{user} in vault"
-      end
-    end
-
-    res = @elastic.request(:put, '/_security/role/nginx', { run_as: actual_users })
-    log.info "update nginx role: #{res}"
-  end
-
-  def sync_ilms
-    log.info 'sync ilm'
-    elastic_ilm = JSON.parse(@elastic.request(:get, '/_ilm/policy'))
-    ilm_for_delete = []
-    elastic_ilm.keys.each do |ilm|
-      # we can't check reserved ilm, so will process only custom ilm starts with twiket
-      # due this any custom ilm must be named with twiket prefix
-      next unless ilm.start_with?('twiket')
-
-      if @config['ilms'][ilm]
-        if @config['ilms'][ilm].key?('retired') && @config['ilms'][ilm]['retired']
-          log.warn "ilm '#{ilm}' retired"
-        else
-          next
-        end
-      end
-      ilm_for_delete << ilm
-    end
-    log.info "will delete ilm: #{ilm_for_delete}"
-
-    ilm_for_create = []
-    @config['ilms'].each do |ilm, params|
-      next if params['retired']
-      ilm_for_create << ilm
-    end
-    log.info "will put ilm: #{ilm_for_create}"
-
-    sleep 3
-
-    ilm_for_delete.each do |ilm|
-      res = @elastic.request(:delete, "/_ilm/policy/#{ilm}")
-      log.info "delete #{ilm}: #{res}"
-    end
-
-    # TODO: (anton.ryabov) mb diff before and put only if changed?
-    # elastic will accept each request and increment policy version even if no difference in policies
-    # how big number they have for version?
-    ilm_for_create.each do |ilm|
-      res = @elastic.request(:put, "/_ilm/policy/#{ilm}", @config['ilms'][ilm]['config'])
-      log.info "create #{ilm}: #{res}"
-    end
-  end
-
-  def sync_templates
-    # 101-200 - 1 shards
-    # 201-300 - 2 shards
-    # 401-500 - 4 shards
-    # 601-700 - 6 shards
-
-    log.info 'sync template'
-    elastic_template = JSON.parse(@elastic.request(:get, '/_template'))
-    template_for_delete = []
-    elastic_template.keys.each do |template|
-      # we can't check reserved template, so will process only custom template starts with twiket
-      # due this any custom template must be named with twiket prefix
-      next unless template.start_with?('twiket')
-
-      if @config['templates'][template]
-        if @config['templates'][template].key?('retired') && @config['templates'][template]['retired']
-          log.warn "template '#{template}' retired"
-        else
-          next
-        end
-      end
-      template_for_delete << template
-    end
-    log.info "will delete template: #{template_for_delete}"
-
-    template_for_create = []
-    @config['templates'].each do |template, params|
-      next if params['retired']
-      template_for_create << template
-    end
-    log.info "will put template: #{template_for_create}"
-
-    sleep 3
-
-    template_for_delete.each do |template|
-      res = @elastic.request(:delete, "/_template/#{template}")
-      log.info "delete #{template}: #{res}"
-    end
-
-    # TODO: (anton.ryabov) mb diff before and put only if changed?
-    # template must have index patterns
-    template_for_create.each do |template|
-      res = @elastic.request(:put, "/_template/#{template}", @config['templates'][template]['config'])
-      log.info "create #{template}: #{res}"
-    end
-  end
-
-  def sync_spaces
-    log.info 'sync spaces'
-    kibana_spaces = JSON.parse(@kibana.request(:get, '/api/spaces/space'))
-    spaces_for_delete = []
-    kibana_spaces.each do |ks|
-      next if ks['_reserved']
-      ks_id = ks['id']
-
-      if @config['spaces'][ks_id]
-        if @config['spaces'][ks_id].key?('retired') && @config['spaces'][ks_id]['retired']
-          log.warn "space '#{ks_id}' retired"
-        else
-          next
-        end
-      end
-      spaces_for_delete << ks_id
-    end
-    log.info "will delete spaces: #{spaces_for_delete}"
-
-    spaces_for_create = []
-    @config['spaces'].each do |space, params|
-      next if params['retired']
-      next if kibana_spaces.map { |ks| ks['id'] }.include?(space)
-      spaces_for_create << space
-    end
-    log.info "will put spaces: #{spaces_for_create}"
-
-    sleep 3
-
-    spaces_for_delete.each do |space|
-      res = @kibana.request(:delete, "/api/spaces/space/#{space}")
-      log.info "delete #{space}: #{res}"
-    end
-
-    spaces_for_create.each do |space|
-      res = @kibana.request(:post, '/api/spaces/space', @config['spaces'][space]['config'])
-      log.info "create #{space}: #{res}"
-    end
   end
 
   def all_indices
